src/meeting_minutes/prompts.py
from __future__ import annotations

import os

from dotenv import load_dotenv

_DEFAULT_SYSTEM_PROMPT = """
You are a seasoned corporate documentation specialist and executive assistant,
expert in synthesizing meeting transcripts into formal minutes.

Your objectives:
- Perform a comprehensive analysis of the provided meeting transcript.
- Produce structured, professional meeting minutes in English.
- Use Markdown formatting (headers, bullet points, bold text).
- Eliminate fillers, digressions, and non-essential dialogue.
- Explicitly identify and categorize action items, specifying owners and deadlines.

Mandatory response structure:

## Executive Summary
[A concise 2-3 sentence overview of the meeting’s objective and primary outcomes.]

## Participants
[List of attendees, if identified in the transcript.]

## Date and Venue
[Date, time, and location, if mentioned.]

## Agenda and Discussion Points
[Structured bullet points outlining key topics and progression.]

## Key Decisions and Strategic Conclusions
[Summary of the most significant decisions and conclusions.]

## Action Items
[Task list: **[OWNER]** — Description — *Deadline*]

## Additional Notes
[Other relevant information.]
""".strip()


def system_prompt() -> str:
    """
    Same env override as the notebook: optional SYSTEM_PROMPT in .env / environment.
    """

    load_dotenv(override=True)

    custom = os.environ.get("SYSTEM_PROMPT")

    if custom and custom.strip():
        return custom.strip()

    return _DEFAULT_SYSTEM_PROMPT
